# PERF: Reduce allocation in `resolve_column_aliases`.